feat(app): update button component to use flowbite-react
This commit updates the button component to use the Flowbite React library. This will improve the look and feel of the button component and make it more consistent with the rest of the application.

feat(lp-items): update lp-items to use flowbite-react

This commit updates the LP_GRID_ITEMS array to use the Flowbite React library. This will improve the look and feel of the LP_GRID_ITEMS array and make it more consistent with the rest of the application.

chore(package.json): update dependencies and package manager

This commit updates the dependencies and package manager to use the latest versions. This will improve the performance and stability of the application.

chore(tailwind.config.js): update tailwind config to use flowbite-react

This commit updates the tailwind config to use the Flowbite React library. This will improve the look and feel of the application and make it more consistent with the rest of the application.